+++ examples/blits-app/src-tauri/src/main.rs
@@ -1,5 +1,0 @@
-#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-
-fn main() {
-    solid_tv_blits_video_example_lib::run();
-}